feat: Improved handling of serial overruns (#31)
Implements a counter in globalState to count the number of serial
overrun errors that we encounter. This can be used to draw out metrics
with Home Assistance or Prometheus. Also tidied up handling so that if
the spi rx method detects a serial overrun, then it should return None
to the caller to be clear that the read can't be trusted.