19 sd card image buildcaptive portal (#45)
Added portal functionality, along with sdcard builder script.

At boot time, a hotspot is created, and presents a captive portal on http://192.168.4.1:80. Most mobile devices should pop up with a captive portal when a device connects to the "openeo" wifi network.

This allows the primary wifi network to be set, a SSH public key to be added to the pi account and for software updates to be retrieved and applied. From the point that the pi connects to an internet connected network, a timer will start and the hotspot will be shut down 30 minutes later for security reasons. 

For the moment, the configuration page remains available on the main LAN on tcp/81, but I suspect we will disable this in due course.