fix: session logger - multiple bugfixes, new features and aesthetic improvements.
Previous versions were quite buggy in how it was recording sessions, so increased accuracy of reported sessions and the charts. The charts now show all of the last N days/weeks/months, rather than just the last N days/weeks/months that there was charging data for, which makes a bit more sense. The charging time is now recorded (this is measured by the number of minutes that the vehicle CT is showing >2A) - this allows us to calculate the accurate average charginging speed for when it was actually charging, rather than just the time that it was connected to the car. For small screens (e.g. phone handset in portrait orientation), extraneous data is hidden in the table to make it readable. On small screens, we also now adjust the font sizes of the charts